Focus overscan in the direction being scrolled to reduce empty space
While a scroll is in progress,  focuses overscan on the direction being scrolled- doubling up the number of overscanned cells.
This reduces the amount of empty space that temporarily appears when a user is quickly scrolling.
It does not increase the overall number of cells being rendered (and so does not negatively impact performance).